requirements.txt
Flask
flask-cors
python-dotenv
gunicorn
langgraph
langchain-groq
langchain-community
duckduckgo-search==5.3.1
APScheduler
python-telegram-bot
